adding fix

+++ debug/rewrite_dataset.py
@@ -74,7 +74,7 @@ def fix_dataset(traj, traj_info):
     traj_pos = tf.cast(traj_pos, tf.float32)
     deltas = tf.linalg.norm(traj_pos[:-1] - traj_pos[1:], axis=-1)
     spacing = tf.reduce_mean(deltas)
-    normalization_factor = lookup_in_dict("normalization_factor", curr_traj_info)
+    normalization_factor = tf.cast(lookup_in_dict("normalization_factor", curr_traj_info), tf.float32)
     tf.print(f"Spacing for {traj_base_name} is {spacing} and normalization factor is {normalization_factor}")
     if tf.abs(spacing - normalization_factor) > 0.05:
         tf.print(f"Spacing issue for {traj_base_name} with spacing {spacing} and normalization factor {normalization_factor}")
